extract some functions to seperated files
